Cover CSV evidence preview
