+++ test/specs/support/by-hand.ts
@@ -8,7 +8,10 @@
 import { expect } from "@std/expect";
 import { leaveEvidencePage } from "#scripts/specs/evidence/pages.ts";
 import { openAdminPage } from "#test/specs/support/browser.ts";
-import { csvDateColumn } from "#test/specs/support/evidence.ts";
+import {
+  csvDateColumn,
+  csvEvidencePage,
+} from "#test/specs/support/evidence.ts";
 import { listingIdNamed } from "#test/specs/support/listings.ts";
 import type { BookingChoices } from "#test/specs/support/public-booking.ts";
 import type {
@@ -21,17 +24,6 @@ import type {
 /** The listing's roster — where the add form and the download both live. */
 const rosterPath = (world: TicketsWorld, name: string): string =>
   `/admin/listing/${listingIdNamed(world, name)}/attendees`;
-
-const escapeHtml = (value: string): string =>
-  value
-    .replaceAll("&", "&amp;")
-    .replaceAll("<", "&lt;")
-    .replaceAll(">", "&gt;");
-
-/** The CSV is the outcome this story reads. Render its date column in the
- * evidence browser so its screenshot proves the exported date range. */
-const csvEvidencePage = (name: string, dates: string): string =>
-  `data:text/html,${encodeURIComponent(`<!doctype html><meta charset="utf-8"><style>:root{--border-radius:8px;--color-accent:#568038;--color-bg:#f1f5ed;--color-bg-secondary:#fff;--color-shadow:#2f4a2024;--color-text:#273320;--font-family:Arial,sans-serif}body{background:var(--color-bg);color:var(--color-text);font:16px var(--font-family);margin:0;padding:32px}main{background:var(--color-bg-secondary);border:1px solid #d7e1cf;border-top:6px solid var(--color-accent);border-radius:var(--border-radius);box-shadow:0 12px 28px var(--color-shadow);margin:auto;max-width:720px;padding:24px}h1{font-size:24px;margin:0 0 20px}pre{background:var(--color-bg);border:1px solid #d7e1cf;overflow-wrap:anywhere;padding:16px;white-space:pre-wrap}</style><main><h1>${escapeHtml(name)} attendee CSV</h1><pre>${escapeHtml(dates)}</pre></main>`)}`;
 
 /** The organiser adds a booking through the form on the listing's roster. Keeps
  * the page they land on, so a story can read what they were told. */
